refactor(binding-mcp-http): move custom json transforms to transform package
Move McpHttpArguments and McpHttpRename from internal.stream to a dedicated
internal.transform package per review feedback; both are common-json
JsonTransform implementations independent of the stream machinery.
